Initial porting to GNU GCC
This is the first version ported to compile and run using the GCC
toolchain provided by the "GNU Tools for ARM Embedded Processors"
project:
https://answers.launchpad.net/gcc-arm-embedded/

This porting has some minor re-arrangements to:
1. convert from windows to unix
2. move some files to have a better and extensible layout
3. make use of a revisited linker-script
4. clean-up some of the major warning
5. add a proper Makefile

The code is still not compiling cleany, but the produced binary image
seems working. It just stops working if you detach the Android device
and then re-connect it.

TODO: clean-up, clean-up and clean-up the code! Warnings are not
acceptable ;-)

Signed-off-by: Patrick Bellasi <derkling@gmail.com>